+++ packages/component-library/components/Dropdown/Dropdown.spec.tsx
@@ -1,4 +1,5 @@
-import { cleanup, fireEvent, render } from "@testing-library/react"
+import { cleanup, render } from "@testing-library/react"
+import { fireEvent } from "@testing-library/dom"
 import * as React from "react"
 import Dropdown from "./Dropdown"