+++ draft-packages/form/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.6.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@2.6.8...@kaizen/draft-form@2.6.9) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
+## [2.6.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@2.6.7...@kaizen/draft-form@2.6.8) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
 ## [2.6.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@2.6.6...@kaizen/draft-form@2.6.7) (2020-09-09)
 
 **Note:** Version bump only for package @kaizen/draft-form