+++ draft-packages/hero-card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@1.4.34...@kaizen/draft-hero-card@1.4.35) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
+## [1.4.34](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@1.4.33...@kaizen/draft-hero-card@1.4.34) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
 ## [1.4.33](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@1.4.32...@kaizen/draft-hero-card@1.4.33) (2020-09-09)
 
 **Note:** Version bump only for package @kaizen/draft-hero-card