+++ draft-packages/radio/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-radio@2.2.20...@kaizen/draft-radio@2.2.21) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-radio
+
+
+
+
+
+## [2.2.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-radio@2.2.19...@kaizen/draft-radio@2.2.20) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-radio
+
+
+
+
+
 ## [2.2.19](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-radio@2.2.18...@kaizen/draft-radio@2.2.19) (2020-09-09)
 
 **Note:** Version bump only for package @kaizen/draft-radio