+++ draft-packages/zen-off-canvas/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.23](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.7.22...@kaizen/draft-zen-off-canvas@1.7.23) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-zen-off-canvas
+
+
+
+
+
+## [1.7.22](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.7.21...@kaizen/draft-zen-off-canvas@1.7.22) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-zen-off-canvas
+
+
+
+
+
 ## [1.7.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-zen-off-canvas@1.7.20...@kaizen/draft-zen-off-canvas@1.7.21) (2020-09-09)
 
 **Note:** Version bump only for package @kaizen/draft-zen-off-canvas