+++ legacy-packages/menu-list/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.33](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.5.32...@kaizen/draft-menu-list@1.5.33) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-menu-list
+
+
+
+
+
+## [1.5.32](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.5.31...@kaizen/draft-menu-list@1.5.32) (2020-09-11)
+
+**Note:** Version bump only for package @kaizen/draft-menu-list
+
+
+
+
+
 ## [1.5.31](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-menu-list@1.5.30...@kaizen/draft-menu-list@1.5.31) (2020-09-09)
 
 **Note:** Version bump only for package @kaizen/draft-menu-list